Remove cx/
This project is now assumed to be a submodule of outherspace-top.git.
cx/ moves to ../cx/, as another submodule of outherspace-top.git.
Likewise, the data directory is assumed to be in ../outherspace-data/...
though the code which assumes this is still assuming it is in data/.
Just specify files manually instead of using the compiled-in test cases.

+++ gui.c
@@ -232,8 +232,8 @@ key_press_fn (Pilot* pilot, RaySpace* space, const SDL_keysym* event)
         if (lights_change > 0)   nphotons += 100;
         else if (nphotons > 10)  nphotons -= 100;
         cast_lights (space, nphotons, nbounces);
-        fprintf (out, "nphotons:%u nlights:%lu\n",
-                 nphotons, (space->lightcuts.nodes.sz+1) / 2);
+        fprintf (out, "nphotons:%u nlights:%u\n",
+                 nphotons, (uint) ((space->lightcuts.nodes.sz+1) / 2));
         nphotons_now = nphotons;
     }
     else if (camera_offset != 0)
